Replaced IGNORESUIT and LEAKYSUIT with leakiness. This allows the "leakiness" UDMF sector property to work.

+++ source/e_udmf.h
@@ -190,8 +190,7 @@ private:
       double rotationfloor = 0.0;
       double rotationceiling = 0.0;
 
-      // TODO: Implement functionality for sliding scale of leakiness from 0-256
-      byte leakiness = 0;
+      int leakiness = 0;
       int damageamount = 0;
       int damageinterval = 0;
       bool damage_endgodmode = false;
